+++ sdk/cosmos/azure-cosmos-encryption/CHANGELOG.md
@@ -1,10 +1,10 @@
 ## Release History
 
-### 1.9.0-beta.1 (Unreleased)
+### 2.0.0-beta.1 (Unreleased)
 
 #### Features Added
-* Added support for allowing parition key path and id to be part of client encryption policy - See [PR 30678](https://github.com/Azure/azure-sdk-for-java/pull/30678)
-
+* Added support for allowing partition key path and id to be part of client encryption policy - See [PR 30678](https://github.com/Azure/azure-sdk-for-java/pull/30678)
+* Added support for allowing the encryption of partition key and id - See [PR 30678](https://github.com/Azure/azure-sdk-for-java/pull/30678)
 
 #### Breaking Changes